requirements.txt
python-dotenv>=1.0.0
redis>=5.0.1
rq>=2.6.0
pymongo>=4.6.1
google-genai>=0.3.0
beautifulsoup4>=4.12.3
fastapi>=0.110.0
uvicorn>=0.29.0
requests>=2.31.0
